+++ cpp/arcae/new_table_proxy.cc
@@ -27,12 +27,15 @@ using ::arrow::Result;
 using ::arrow::Status;
 using ::arrow::Table;
 
+using LockType = ::casacore::FileLocker::LockType;
 using ::casacore::JsonOut;
 using ::casacore::JsonParser;
 using ::casacore::Record;
 using ::casacore::TableProxy;
 
 namespace arcae {
+
+// Table Read Operations
 
 Result<std::string> NewTableProxy::GetTableDescriptor() const {
   return itp_
@@ -112,13 +115,6 @@ Result<std::shared_ptr<Array>> NewTableProxy::GetRowShapes(
       .MoveResult();
 }
 
-Result<bool> NewTableProxy::PutColumn(const std::string& column,
-                                      const std::shared_ptr<Array>& data,
-                                      const detail::Selection& selection) const {
-  ARROW_RETURN_NOT_OK(SafeMultithreadedWrites());
-  return WriteImpl(itp_, column, data, selection).MoveResult();
-}
-
 Result<std::string> NewTableProxy::Name() const {
   return itp_
       ->RunAsync(
@@ -149,38 +145,42 @@ Result<std::size_t> NewTableProxy::nRows() const {
       .MoveResult();
 }
 
+// Table Write Operations from this point onwards
+
 Result<bool> NewTableProxy::AddRows(std::size_t nrows) {
-  ARROW_RETURN_NOT_OK(SafeMultithreadedWrites());
-  return itp_
-      ->RunAsync([nrows = nrows](TableProxy& tp) {
-        detail::MaybeReopenRW(tp);
-        tp.addRow(nrows);
-        return true;
-      })
+  return itp_->SpawnWriter()
+      ->RunAsync(
+          [nrows = nrows](TableProxy& tp) {
+            detail::MaybeReopenRW(tp);
+            tp.addRow(nrows);
+            return true;
+          },
+          LockType::Write)
       .MoveResult();
 }
 
 Result<bool> NewTableProxy::AddColumns(const std::string& json_columndescs,
                                        const std::string& json_dminfo) {
-  ARROW_RETURN_NOT_OK(SafeMultithreadedWrites());
-  return itp_
-      ->RunAsync([json_columndescs = json_columndescs,
-                  json_dminfo = json_dminfo](TableProxy& tp) {
-        detail::MaybeReopenRW(tp);
-        Record columndescs = JsonParser::parse(json_columndescs).toRecord();
-        Record dminfo = JsonParser::parse(json_dminfo).toRecord();
-        tp.addColumns(columndescs, dminfo, false);
-        return true;
-      })
+  return itp_->SpawnWriter()
+      ->RunAsync(
+          [json_columndescs = json_columndescs,
+           json_dminfo = json_dminfo](TableProxy& tp) {
+            detail::MaybeReopenRW(tp);
+            Record columndescs = JsonParser::parse(json_columndescs).toRecord();
+            Record dminfo = JsonParser::parse(json_dminfo).toRecord();
+            tp.addColumns(columndescs, dminfo, false);
+            return true;
+          },
+          LockType::Write)
       .MoveResult();
 }
 
-Result<bool> NewTableProxy::Close() { return itp_->Close(); }
-
-Status NewTableProxy::SafeMultithreadedWrites() const {
-  if (itp_->nInstances() == 1) return Status::OK();
-  return Status::NotImplemented("Write support when number of table instances ",
-                                itp_->nInstances(), " is greater than one");
+Result<bool> NewTableProxy::PutColumn(const std::string& column,
+                                      const std::shared_ptr<Array>& data,
+                                      const detail::Selection& selection) const {
+  return WriteImpl(itp_->SpawnWriter(), column, data, selection).MoveResult();
 }
+
+Result<bool> NewTableProxy::Close() { return itp_->Close(); }
 
 }  // namespace arcae